requirements.txt
crewai
langchain
langchain-openai
streamlit
python-dotenv
requests
pandas
openai